commit success get data

+++ core/src/main/java/com/example/core/data/source/remote/response/CoinInfo.kt
@@ -13,7 +13,7 @@ data class CoinInfo(
     val Internal: String,
     val MaxSupply: Double,
     val Name: String,
-    val NetHashesPerSecond: Long,
+    val NetHashesPerSecond: Double,
     val ProofType: String,
     val Rating: Rating,
     val Type: Int,
