+++ core/src/main/java/com/example/core/data/source/remote/response/USDX.kt
@@ -31,7 +31,7 @@ data class USDX(
     val OPENDAY: Double,
     val OPENHOUR: Double,
     val PRICE: Double,
-    val SUPPLY: Int,
+    val SUPPLY: Double,
     val TOPTIERVOLUME24HOUR: Double,
     val TOPTIERVOLUME24HOURTO: Double,
     val TOSYMBOL: String,
